feat: Write a type-annotated function to_kv that takes a string k and an int OR float v as arguments and returns a tuple.

+++ python_variable_annotations/7-to_kv.py
@@ -5,6 +5,6 @@ and returns a tuple"""
 from typing import Tuple, Union
 
 
-def to_kv(k: str, v: Union[int, float]) -> Tuple[str, int]:
+def to_kv(k: str, v: Union[int, float]) -> Tuple[str, float]:
     """returns a tuple of the string k and the square of v"""
     return (k, v ** 2)
